+++ lib/src/core/engine.dart
@@ -16,27 +16,38 @@
 
 import 'dart:async';
 
-import 'package:flutter/foundation.dart';
+import 'package:flutter/foundation.dart' hide internal;
 
 import 'package:collection/collection.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter_webrtc/flutter_webrtc.dart' as rtc;
 import 'package:meta/meta.dart';
 
-import 'package:livekit_client/livekit_client.dart';
+import '../events.dart';
+import '../exceptions.dart';
 import '../extensions.dart';
 import '../internal/events.dart';
 import '../internal/types.dart';
+import '../logger.dart' show logger;
+import '../managers/event.dart';
+import '../options.dart';
 import '../proto/livekit_models.pb.dart' as lk_models;
 import '../proto/livekit_rtc.pb.dart' as lk_rtc;
+import '../publication/local.dart';
 import '../support/disposable.dart';
 import '../support/region_url_provider.dart';
 import '../support/websocket.dart';
+import '../track/local/local.dart';
+import '../track/local/video.dart';
 import '../types/internal.dart';
 import '../utils/data_packet_buffer.dart';
 import '../utils/ttl_map.dart';
+import '../types/other.dart';
 import 'signal_client.dart';
 import 'transport.dart';
+
+import '../support/platform.dart'
+    show lkPlatformIsTest, lkPlatformIs, PlatformType;
 
 const maxRetryDelay = 7000;
 
@@ -312,7 +323,11 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
     if (isBufferStatusLow(kind) == true) {
       completer.complete();
     } else {
-      onClosing() => completer.completeError('Engine disconnected');
+      onClosing() {
+        if (!completer.isCompleted) {
+          completer.completeError('Engine disconnected');
+        }
+      }
       events.once<EngineClosingEvent>((e) => onClosing());
 
       while (!_dcBufferStatus[kind]!) {
@@ -364,7 +379,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
     lk_models.DataPacket packet, {
     bool? reliability = true,
   }) async {
-    var reliabilityType =
+    final reliabilityType =
         reliability == true ? Reliability.reliable : Reliability.lossy;
 
     // Block sending during reconnection attempts for reliable packets
@@ -682,7 +697,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
   Future<void> _handleGettingConnectedServerAddress(
       rtc.RTCPeerConnection pc) async {
     try {
-      var remoteAddress = await getConnectedAddress(publisher!.pc);
+      final remoteAddress = await getConnectedAddress(publisher!.pc);
       logger.fine('Connected address: $remoteAddress');
       if (_connectedServerAddress == null ||
           _connectedServerAddress != remoteAddress) {
@@ -803,7 +818,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
     _isReconnecting = true;
 
     if (reconnectAttempts == 0) {
-      reconnectStart = DateTime.now();
+      reconnectStart = DateTime.timestamp();
     }
 
     if (reconnectAttempts! >= _reconnectCount) {
@@ -817,7 +832,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
       return;
     }
 
-    var delay = defaultRetryDelaysInMs[reconnectAttempts!];
+    final delay = defaultRetryDelaysInMs[reconnectAttempts!];
 
     events.emit(EngineAttemptReconnectEvent(
       attempt: reconnectAttempts! + 1,
@@ -1058,7 +1073,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
       // create peer connections
       _subscriberPrimary = event.response.subscriberPrimary;
       _serverInfo = event.response.serverInfo;
-      var iceServersFromServer =
+      final iceServersFromServer =
           event.response.iceServers.map((e) => e.toSDKType()).toList();
 
       if (iceServersFromServer.isNotEmpty) {
@@ -1072,7 +1087,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
           'iceServers: ${event.response.iceServers}, '
           'forceRelay: $event.response.clientConfiguration.forceRelay');
 
-      var rtcConfiguration = await _buildRtcConfiguration(
+      final rtcConfiguration = await _buildRtcConfiguration(
           serverResponseForceRelay:
               event.response.clientConfiguration.forceRelay,
           serverProvidedIceServers: _serverProvidedIceServers);
@@ -1093,7 +1108,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
       events.emit(EngineJoinResponseEvent(response: event.response));
     })
     ..on<SignalReconnectResponseEvent>((event) async {
-      var iceServersFromServer =
+      final iceServersFromServer =
           event.response.iceServers.map((e) => e.toSDKType()).toList();
 
       if (iceServersFromServer.isNotEmpty) {
@@ -1107,7 +1122,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
           'forceRelay: $event.response.clientConfiguration.forceRelay, '
           'lastMessageSeq: ${event.response.lastMessageSeq}');
 
-      var rtcConfiguration = await _buildRtcConfiguration(
+      final rtcConfiguration = await _buildRtcConfiguration(
           serverResponseForceRelay:
               event.response.clientConfiguration.forceRelay,
           serverProvidedIceServers: _serverProvidedIceServers);
@@ -1154,7 +1169,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
         logger.warning('[$objectId] subscriber is null');
         return;
       }
-      var signalingState = await subscriber!.pc.getSignalingState();
+      final signalingState = await subscriber!.pc.getSignalingState();
       logger.fine('[$objectId] Received server offer(type: ${event.sd.type}, '
           '$signalingState)');
       logger.finer('sdp: ${event.sd.sdp}');
@@ -1281,7 +1296,7 @@ extension EngineInternalMethods on Engine {
     if (track.mediaStreamTrack.kind == 'video' && opts is VideoPublishOptions) {
       track.codec = opts.videoCodec;
     }
-    var transceiverInit = rtc.RTCRtpTransceiverInit(
+    final transceiverInit = rtc.RTCRtpTransceiverInit(
       direction: rtc.TransceiverDirection.SendOnly,
     );
     if (encodings != null) {
@@ -1314,7 +1329,7 @@ extension EngineInternalMethods on Engine {
     if (publisher == null) {
       throw Exception('publisher is closed');
     }
-    var transceiverInit = rtc.RTCRtpTransceiverInit(
+    final transceiverInit = rtc.RTCRtpTransceiverInit(
       direction: rtc.TransceiverDirection.SendOnly,
     );
     if (encodings != null) {
@@ -1334,22 +1349,22 @@ extension EngineInternalMethods on Engine {
       rtc.RTCRtpTransceiver transceiver, String kind, String videoCodec) async {
     // when setting codec preferences, the capabilites need to be read from
     // the RTCRtpReceiver
-    var caps = await rtc.getRtpReceiverCapabilities(kind);
+    final caps = await rtc.getRtpReceiverCapabilities(kind);
     if (caps.codecs == null) return;
 
     logger.fine('get capabilities ${caps.codecs}');
 
-    List<rtc.RTCRtpCodecCapability> matched = [];
-    List<rtc.RTCRtpCodecCapability> partialMatched = [];
-    List<rtc.RTCRtpCodecCapability> unmatched = [];
+    final List<rtc.RTCRtpCodecCapability> matched = [];
+    final List<rtc.RTCRtpCodecCapability> partialMatched = [];
+    final List<rtc.RTCRtpCodecCapability> unmatched = [];
     for (var c in caps.codecs!) {
-      var codec = c.mimeType.toLowerCase();
+      final codec = c.mimeType.toLowerCase();
       if (codec == 'audio/opus') {
         matched.add(c);
         continue;
       }
 
-      var matchesVideoCodec = codec == 'video/$videoCodec';
+      final matchesVideoCodec = codec == 'video/$videoCodec';
       if (!matchesVideoCodec) {
         if (lkPlatformIs(PlatformType.android) && codec == 'video/vp9') {
           if (c.sdpFmtpLine != null &&