+++ lib/src/support/region_url_provider.dart
@@ -23,7 +23,7 @@ class RegionUrlProvider {
   RegionUrlProvider({required String url, required this.token})
       : serverUrl = Uri.parse(url);
 
-  updateToken(String token) {
+  void updateToken(String token) {
     this.token = token;
   }
 
@@ -39,7 +39,7 @@ class RegionUrlProvider {
           'region availability is only supported for LiveKit Cloud domains');
     }
     if (regionSettings == null ||
-        DateTime.now().microsecondsSinceEpoch - lastUpdateAt >
+        DateTime.timestamp().microsecondsSinceEpoch - lastUpdateAt >
             settingsCacheTime) {
       regionSettings = await fetchRegionSettings();
     }
@@ -57,29 +57,29 @@ class RegionUrlProvider {
     }
   }
 
-  resetAttempts() {
+  void resetAttempts() {
     attemptedRegions = [];
   }
 
   /* @internal */
   Future<lk_models.RegionSettings> fetchRegionSettings() async {
-    var url = '${getCloudConfigUrl(serverUrl)}/regions';
-    http.Response regionSettingsResponse =
+    final url = '${getCloudConfigUrl(serverUrl)}/regions';
+    final http.Response regionSettingsResponse =
         await http.get(Uri.parse(url), headers: {
       'authorization': 'Bearer $token',
     });
     if (regionSettingsResponse.statusCode == 200) {
-      var mapData = json.decode(regionSettingsResponse.body);
-      var regions = (mapData['regions'] as List<dynamic>)
+      final mapData = json.decode(regionSettingsResponse.body);
+      final regions = (mapData['regions'] as List<dynamic>)
           .map((region) => lk_models.RegionInfo(
               distance: Int64(int.parse(region['distance'])),
               region: region['region'],
               url: region['url']))
           .toList();
-      var regionSettings = lk_models.RegionSettings(
+      final regionSettings = lk_models.RegionSettings(
         regions: regions,
       );
-      lastUpdateAt = DateTime.now().microsecondsSinceEpoch;
+      lastUpdateAt = DateTime.timestamp().microsecondsSinceEpoch;
       return regionSettings;
     } else {
       throw ConnectException(
@@ -91,9 +91,9 @@ class RegionUrlProvider {
     }
   }
 
-  setServerReportedRegions(lk_models.RegionSettings regions) {
+  void setServerReportedRegions(lk_models.RegionSettings regions) {
     regionSettings = regions;
-    lastUpdateAt = DateTime.now().millisecondsSinceEpoch;
+    lastUpdateAt = DateTime.timestamp().millisecondsSinceEpoch;
   }
 
   String getCloudConfigUrl(Uri serverUrl) {