+++ lib/src/track/audio_visualizer.dart
@@ -1,8 +1,9 @@
-import 'package:livekit_client/src/support/disposable.dart';
+import 'package:uuid/uuid.dart' as uuid;
+
 import '../events.dart' show AudioVisualizerEvent;
 import '../managers/event.dart' show EventsEmittable;
+import '../support/disposable.dart';
 import 'local/local.dart' show AudioTrack;
-import 'package:uuid/uuid.dart' as uuid;
 
 import 'audio_visualizer_native.dart'
     if (dart.library.js_interop) 'audio_visualizer_web.dart';