+++ lib/src/track/audio_visualizer_native.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/services.dart';
+
 import 'package:flutter_webrtc/flutter_webrtc.dart';
 
 import '../events.dart' show AudioVisualizerEvent;