+++ lib/src/track/local/video.dart
@@ -17,9 +17,9 @@ import 'package:flutter/foundation.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter_webrtc/flutter_webrtc.dart' as rtc;
 
-import 'package:livekit_client/src/extensions.dart';
 import '../../events.dart';
 import '../../exceptions.dart';
+import '../../extensions.dart';
 import '../../logger.dart';
 import '../../options.dart';
 import '../../proto/livekit_models.pb.dart' as lk_models;
@@ -40,11 +40,12 @@ class SimulcastTrackInfo {
 
   List<rtc.RTCRtpEncoding>? encodings;
 
-  SimulcastTrackInfo(
-      {required this.codec,
-      this.encodings,
-      required this.mediaStreamTrack,
-      this.sender});
+  SimulcastTrackInfo({
+    required this.codec,
+    this.encodings,
+    required this.mediaStreamTrack,
+    this.sender,
+  });
 }
 
 /// A video track from the local device. Use static methods in this class to create
@@ -57,11 +58,13 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
   VideoPublishOptions? lastPublishOptions;
 
   num? _currentBitrate;
-  get currentBitrate => _currentBitrate;
+  num? get currentBitrate => _currentBitrate;
+
   Map<String, VideoSenderStats>? prevStats;
   final Map<String, num> _bitrateFoLayers = {};
 
   Map<String, SimulcastTrackInfo> simulcastCodecs = {};
+  Map<(String, int), rtc.RTCRtpEncoding> encodingBackups = {};
 
   List<lk_rtc.SubscribedCodec> subscribedCodecs = [];
 
@@ -78,7 +81,7 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
       logger.warning('Failed to get sender stats: $e');
       return false;
     }
-    Map<String, VideoSenderStats> statsMap = {};
+    final Map<String, VideoSenderStats> statsMap = {};
 
     for (var s in stats) {
       statsMap[s.rid ?? 'f'] = s;
@@ -92,7 +95,7 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
           return;
         }
         try {
-          var bitRateForlayer = computeBitrateForSenderStats(s, prev).toInt();
+          final bitRateForlayer = computeBitrateForSenderStats(s, prev).toInt();
           _bitrateFoLayers[key] = bitRateForlayer;
           totalBitrate += bitRateForlayer;
         } catch (e) {
@@ -100,11 +103,13 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
         }
       });
       _currentBitrate = totalBitrate;
-      events.emit(VideoSenderStatsEvent(
-        stats: statsMap,
-        currentBitrate: currentBitrate,
-        bitrateForLayers: _bitrateFoLayers,
-      ));
+      events.emit(
+        VideoSenderStatsEvent(
+          stats: statsMap,
+          currentBitrate: totalBitrate,
+          bitrateForLayers: _bitrateFoLayers,
+        ),
+      );
     }
 
     prevStats = statsMap;
@@ -117,10 +122,10 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
     }
 
     final stats = await sender!.getStats();
-    List<VideoSenderStats> items = [];
+    final List<VideoSenderStats> items = [];
     for (var v in stats) {
       if (v.type == 'outbound-rtp') {
-        VideoSenderStats vs = VideoSenderStats(v.id, v.timestamp);
+        final vs = VideoSenderStats(v.id, v.timestamp);
         vs.frameHeight = getNumValFromReport(v.values, 'frameHeight');
         vs.frameWidth = getNumValFromReport(v.values, 'frameWidth');
         vs.framesPerSecond = getNumValFromReport(v.values, 'framesPerSecond');
@@ -131,14 +136,22 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
         vs.bytesSent = getNumValFromReport(v.values, 'bytesSent');
         vs.framesSent = getNumValFromReport(v.values, 'framesSent');
         vs.rid = getStringValFromReport(v.values, 'rid');
-        vs.encoderImplementation =
-            getStringValFromReport(v.values, 'encoderImplementation');
-        vs.retransmittedPacketsSent =
-            getNumValFromReport(v.values, 'retransmittedPacketsSent');
-        vs.qualityLimitationReason =
-            getStringValFromReport(v.values, 'qualityLimitationReason');
-        vs.qualityLimitationResolutionChanges =
-            getNumValFromReport(v.values, 'qualityLimitationResolutionChanges');
+        vs.encoderImplementation = getStringValFromReport(
+          v.values,
+          'encoderImplementation',
+        );
+        vs.retransmittedPacketsSent = getNumValFromReport(
+          v.values,
+          'retransmittedPacketsSent',
+        );
+        vs.qualityLimitationReason = getStringValFromReport(
+          v.values,
+          'qualityLimitationReason',
+        );
+        vs.qualityLimitationResolutionChanges = getNumValFromReport(
+          v.values,
+          'qualityLimitationResolutionChanges',
+        );
 
         // locate the appropriate remote-inbound-rtp item
         final remoteId = getStringValFromReport(v.values, 'remoteId');
@@ -162,14 +175,12 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
   }
 
   // Private constructor
-  LocalVideoTrack._(TrackSource source, rtc.MediaStream stream,
-      rtc.MediaStreamTrack track, this.currentOptions)
-      : super(
-          TrackType.VIDEO,
-          source,
-          stream,
-          track,
-        );
+  LocalVideoTrack._(
+    TrackSource source,
+    rtc.MediaStream stream,
+    rtc.MediaStreamTrack track,
+    this.currentOptions,
+  ) : super(TrackType.VIDEO, source, stream, track);
 
   /// Creates a LocalVideoTrack from camera input.
   static Future<LocalVideoTrack> createCameraTrack([
@@ -178,7 +189,7 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
     options ??= const CameraCaptureOptions();
 
     final stream = await LocalTrack.createStream(options);
-    var track = LocalVideoTrack._(
+    final track = LocalVideoTrack._(
       TrackSource.camera,
       stream,
       stream.getVideoTracks().first,
@@ -201,7 +212,8 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
   ]) async {
     if (lkPlatformIsWebMobile()) {
       throw TrackCreateException(
-          'Screen sharing is not supported on mobile devices');
+        'Screen sharing is not supported on mobile devices',
+      );
     }
     options ??= const ScreenShareCaptureOptions();
 
@@ -224,7 +236,8 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
   ]) async {
     if (lkPlatformIsWebMobile()) {
       throw TrackCreateException(
-          'Screen sharing is not supported on mobile devices');
+        'Screen sharing is not supported on mobile devices',
+      );
     }
     if (options == null) {
       options = const ScreenShareCaptureOptions(captureScreenAudio: true);
@@ -233,18 +246,24 @@ class LocalVideoTrack extends LocalTrack with VideoTrack {
     }
     final stream = await LocalTrack.createStream(options);
 
-    List<LocalTrack> tracks = [
+    final List<LocalTrack> tracks = [
       LocalVideoTrack._(
         TrackSource.screenShareVideo,
         stream,
         stream.getVideoTracks().first,
         options,
-      )
+      ),
     ];
 
     if (stream.getAudioTracks().isNotEmpty) {
-      tracks.add(LocalAudioTrack(TrackSource.screenShareAudio, stream,
-          stream.getAudioTracks().first, const AudioCaptureOptions()));
+      tracks.add(
+        LocalAudioTrack(
+          TrackSource.screenShareAudio,
+          stream,
+          stream.getAudioTracks().first,
+          const AudioCaptureOptions(),
+        ),
+      );
     }
     return tracks;
   }
@@ -262,10 +281,11 @@ extension LocalVideoTrackExt on LocalVideoTrack {
       return;
     }
     final newOptions = CameraCaptureOptions(
-        cameraPosition: position,
-        deviceId: null,
-        maxFrameRate: options.maxFrameRate,
-        params: options.params);
+      cameraPosition: position,
+      deviceId: null,
+      maxFrameRate: options.maxFrameRate,
+      params: options.params,
+    );
     await restartTrack(newOptions);
     await replaceTrackForMultiCodecSimulcast(mediaStreamTrack);
     currentOptions = newOptions;
@@ -284,15 +304,14 @@ extension LocalVideoTrackExt on LocalVideoTrack {
       return;
     }
 
-    await restartTrack(
-      options.copyWith(deviceId: deviceId),
-    );
+    await restartTrack(options.copyWith(deviceId: deviceId));
 
     await replaceTrackForMultiCodecSimulcast(mediaStreamTrack);
   }
 
   Future<void> replaceTrackForMultiCodecSimulcast(
-      rtc.MediaStreamTrack newTrack) async {
+    rtc.MediaStreamTrack newTrack,
+  ) async {
     simulcastCodecs.forEach((key, simulcastTrack) async {
       await simulcastTrack.sender?.replaceTrack(newTrack);
       simulcastTrack.mediaStreamTrack = mediaStreamTrack;
@@ -300,7 +319,9 @@ extension LocalVideoTrackExt on LocalVideoTrack {
   }
 
   Future<List<String>> setPublishingCodecs(
-      List<lk_rtc.SubscribedCodec> codecs, LocalTrack track) async {
+    List<lk_rtc.SubscribedCodec> codecs,
+    LocalTrack track,
+  ) async {
     logger.fine('setPublishingCodecs $codecs');
 
     // only enable simulcast codec for preference codec setted
@@ -311,7 +332,7 @@ extension LocalVideoTrackExt on LocalVideoTrack {
 
     subscribedCodecs = codecs;
 
-    List<String> newCodecs = [];
+    final List<String> newCodecs = [];
 
     for (var codec in codecs) {
       if (this.codec?.toLowerCase() == codec.codec.toLowerCase()) {
@@ -341,7 +362,9 @@ extension LocalVideoTrackExt on LocalVideoTrack {
   }
 
   Future<void> updatePublishingLayers(
-      LocalTrack? track, List<lk_rtc.SubscribedQuality> layers) async {
+    LocalTrack? track,
+    List<lk_rtc.SubscribedQuality> layers,
+  ) async {
     logger.fine('Update publishing layers: $layers');
 
     if (track?.sender == null) {
@@ -378,9 +401,10 @@ extension LocalVideoTrackExt on LocalVideoTrack {
   }
 
   Future<void> setPublishingLayersForSender(
-      rtc.RTCRtpSender sender,
-      List<rtc.RTCRtpEncoding> encodings,
-      List<lk_rtc.SubscribedQuality> layers) async {
+    rtc.RTCRtpSender sender,
+    List<rtc.RTCRtpEncoding> encodings,
+    List<lk_rtc.SubscribedQuality> layers,
+  ) async {
     logger.fine('Update publishing layers: $layers');
 
     final params = sender.parameters;
@@ -395,7 +419,7 @@ extension LocalVideoTrackExt on LocalVideoTrack {
     /* @ts-ignore */
     if (encodings[0].scalabilityMode != null) {
       // svc dynacast encodings
-      var encoding = encodings[0];
+      final encoding = encodings[0];
       /* @ts-ignore */
       // const mode = new ScalabilityMode(encoding.scalabilityMode);
       var maxQuality = lk_models.VideoQuality.OFF;
@@ -435,9 +459,10 @@ extension LocalVideoTrackExt on LocalVideoTrack {
         if (rid == '') {
           rid = 'q';
         }
-        var quality = _videoQualityForRid(rid);
-        var subscribedQuality =
-            layers.firstWhereOrNull((q) => q.quality == quality);
+        final quality = _videoQualityForRid(rid);
+        final subscribedQuality = layers.firstWhereOrNull(
+          (q) => q.quality == quality,
+        );
         if (subscribedQuality == null) {
           continue;
         }
@@ -452,11 +477,18 @@ extension LocalVideoTrackExt on LocalVideoTrack {
           // have a workaround of lowering its bitrate and resolution to the min.
           if (kIsWeb && lkBrowser() == BrowserType.firefox) {
             if (subscribedQuality.enabled) {
+              final encodingBackup =
+                  encodingBackups[(sender.senderId, idx)] ?? encoding;
               encoding.scaleResolutionDownBy =
-                  encodings[idx].scaleResolutionDownBy;
-              encoding.maxBitrate = encodings[idx].maxBitrate;
-              encoding.maxFramerate = encodings[idx].maxBitrate;
+                  encodingBackup.scaleResolutionDownBy;
+              encoding.maxBitrate = encodingBackup.maxBitrate;
+              encoding.maxFramerate = encodingBackup.maxFramerate;
             } else {
+              encodingBackups[(sender.senderId, idx)] = rtc.RTCRtpEncoding(
+                scaleResolutionDownBy: encoding.scaleResolutionDownBy,
+                maxBitrate: encoding.maxBitrate,
+                maxFramerate: encoding.maxFramerate,
+              );
               encoding.scaleResolutionDownBy = 4;
               encoding.maxBitrate = 10;
               encoding.maxFramerate = 2;
@@ -481,11 +513,13 @@ extension LocalVideoTrackExt on LocalVideoTrack {
   }
 
   SimulcastTrackInfo addSimulcastTrack(
-      String codec, List<rtc.RTCRtpEncoding> encodings) {
+    String codec,
+    List<rtc.RTCRtpEncoding> encodings,
+  ) {
     if (simulcastCodecs[codec] != null) {
       throw Exception('$codec already added');
     }
-    SimulcastTrackInfo simulcastCodecInfo = SimulcastTrackInfo(
+    final SimulcastTrackInfo simulcastCodecInfo = SimulcastTrackInfo(
       codec: codec,
       encodings: encodings,
       mediaStreamTrack: mediaStreamTrack,