+++ lib/src/widgets/screen_select_dialog.dart
@@ -152,7 +152,7 @@ class ScreenSelectDialog extends Dialog {
 
   Future<void> _getSources() async {
     try {
-      var sources = await rtc.desktopCapturer.getSources(types: [_sourceType]);
+      final sources = await rtc.desktopCapturer.getSources(types: [_sourceType]);
       for (var element in sources) {
         if (kDebugMode) {
           print(